More accurate offline detection
Instead of incrementing and decrementing a counter with each ping or
"check up" expiration, fork a thread to patiently mark us offline if
we don't receive a heartbeat when expected.

This eliminates sleep polling and makes it easier to see that we
become offline with 'intervalDead' seconds of no heartbeat.