Remove redundant 'online' field
Because we only track ping counts when online and those values are
always positive, we can use a negative ping count to track online
status.  This is a bit of a hack, but it seems like the best way to
capture all the relevant invariants.  I tried defining

    data Ekg = Online  { thread::ThreadId, pings::Int }
             | Offline { thread::ThreadId }

but that made the rest of the code harder for me to read.  If this were a
hugely vital production system, the above would be worth doing to move some
invariant checking to compile time.